minor improvements

+++ src/hll/hll.h
@@ -14,19 +14,19 @@ struct HLL {
   uint8_t bits;
   uint8_t rank_bits;  // cached value of 32 - bits. Represents the number of bits used for the rank/max rank
   uint32_t size;      // number of registers (2^bits). bits <= 20 so this fits in 32 bits
-  double cachedCard;  // cached cardinality. -1 if not cached
+  size_t cachedCard;  // cached cardinality
   uint8_t *registers;
 };
 
 extern int hll_init(struct HLL *hll, uint8_t bits);
-extern int hll_load(struct HLL *hll, const void *registers, size_t size);
+extern int hll_load(struct HLL *hll, const void *registers, uint32_t size);
 extern void hll_destroy(struct HLL *hll);
 extern int hll_merge(struct HLL *dst, const struct HLL *src);
 extern void hll_add(struct HLL *hll, const void *buf, size_t size);
 void hll_add_hash(struct HLL *hll, uint32_t h);
-extern double hll_count(const struct HLL *hll);
-int hll_set_registers(struct HLL *hll, const void *registers, size_t size);
-int hll_merge_registers(struct HLL *hll, const void *registers, size_t size);
+extern size_t hll_count(const struct HLL *hll);
+int hll_set_registers(struct HLL *hll, const void *registers, uint32_t size);
+int hll_merge_registers(struct HLL *hll, const void *registers, uint32_t size);
 void hll_clear(struct HLL *hll);
 
 #endif /* AVZ_HLL_H */
